refactor(core): decompose country_service_registry — data rows out of the registry (#3232)
The registry was a 864-line god-file (6 grandfathered file_length bumps,
decomposition forced by #3141): the per-service FuelServicePolicy consts and
the ordered CountryServiceEntry rows — pure data — sat alongside the lookups
and service builders. Split the data out so the registry holds only behaviour.

- country_service_entry.dart (90): the CountryServiceEntry model. Re-exported
  from country_service_registry.dart so the ~30 sites that import the type via
  the registry keep resolving it unchanged.
- country_service_policies.dart (277): the per-service FuelServicePolicy consts
  (now public k…Policy; the staged legacy/bulk variants stay private).
- country_service_data.dart (316): kDefaultFuelTypes + the ordered
  kCountryServiceEntries list.
- country_service_registry.dart (864 → 241): lookups + foreground/background
  service builders only. `entries` is now a const alias of kCountryServiceEntries,
  so every CountryServiceRegistry.entries call site is unchanged.

Pure relocation — identical data, identical behaviour. Removed the registry's
grandfather entry from the file_length snapshot (all four files < 400 cap).
Repointed germany_no_special_case_test's entry-existence assertions at the data
file.
